# swagger/swagger.yaml
...
swagger: "2.0"
...
host: hogehoge.luckylife.cloud:3000
...
info:
  title: テスト用 GUI
  description: Design-based Management Console. Presented by Viron
  version: 0.1.0

#################################################
# schemes
# - APIを提供する通信プロトコルのリスト
#################################################
schemes:
- https

#################################################
# definitions
# - リクエストやレスポンスのレイアウトの定義
# - paths内から $ref で参照できる
#################################################
definitions:
...
  Error:
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: #F-0000-0001
      meta:
        timestamp: 1502172152
      status: "400"
    properties:
      code:
        description: エラーコード
        type: string
      detail:
        description: 詳細
        type: string
      id:
        description: ID
        type: string
      meta:
        additionalProperties: true
        description: メタ情報
        type: object
      status:
        description: ステータスコード
        type: string
    type: object
...
paths:
...
  /id_masters/download/csv:
    get:
      x-swagger-router-controller: idmasters
      description: 割当マスタ一覧をcsvファイルでダウンロード
      operationId: idmasters#download
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - jwt:
        - api:access
      summary: 割当マスタ一覧ダウンロード
      x-submit-label: 割当マスタ一覧をダウンロードする
      tags:
      - idmasters
...
  /divisions/download/csv:
    get:
      x-swagger-router-controller: divisions
      description: 組織一覧をcsvファイルでダウンロード
      operationId: divisions#download
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - jwt:
        - api:access
      summary: 組織一覧ダウンロード
      x-submit-label: 組織一覧をダウンロードする
      tags:
      - divisions
...
  /user/download/csv:
    get:
      x-swagger-router-controller: user
      description: ユーザー一覧をcsvファイルでダウンロード
      operationId: user#download
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - jwt:
        - api:access
      summary: ユーザー一覧ダウンロード
      x-submit-label: ユーザー一覧をダウンロードする
      tags:
      - user
...
securityDefinitions:
  jwt:
    type: oauth2
    authorizationUrl: https://localhost:3000/signin
    flow: implicit
    scopes:
      api:access: API access

#################################################
# produces
# - APIが提供するMIMEタイプの一覧
#################################################
produces:
  - application/json
  - text/plain

#################################################
# consumes
# - APIが受け付けるMIMEタイプの一覧
#################################################
consumes:
  - application/json
  - text/plain